+++ cryptography/exceptions.py
@@ -34,3 +34,11 @@ class InvalidTag(Exception):
 
 class InvalidSignature(Exception):
     pass
+
+
+class InternalError(Exception):
+    pass
+
+
+class InvalidKey(Exception):
+    pass